Refactoring of NLG metrics
